fix bug

+++ ohos_adapter/hiviewdfx_adapter/include/hisysevent_adapter_impl.h
@@ -56,7 +56,7 @@ public:
     int Write(const std::string& eventName, EventType type,
         const std::tuple<const std::string, const int64_t, const std::string, const int64_t,
         const std::string, const int, const std::string, const int,
-        const std::string, const int64_t, const std::string, const int64_t>& data) override;
+        const std::string, const int64_t, const std::string, const int>& data) override;
 };
 } // namespace OHOS::NWeb
